+++ Backend/src/app/controllers/productController.ts
@@ -1,7 +1,8 @@
 /* eslint-disable consistent-return */
 /* eslint-disable no-unused-vars */
 import { Request, Response, NextFunction } from 'express';
-import { IProducts, Product } from '../models/product';
+import { Product } from '../models/product';
+import { IProducts } from '../interface/Iproducts';
 import { Error_Handler } from '../utils/errorHandling';
 import asyncError from '../middlewares/asyncError';
 import { Api_Features } from '../utils/apiFeatures';
@@ -9,9 +10,7 @@ import { Api_Features } from '../utils/apiFeatures';
 export class Product_Controllers {
   public createProducts = asyncError(async (req: Request, res: Response, _next: NextFunction): Promise<void> => {
     req.body.user = req.body.id;
-
     const product = await Product.create(req.body);
-
     res.status(201).json({
       success: true,
       product,
@@ -20,22 +19,15 @@ export class Product_Controllers {
 
   public getProducts = asyncError(async (req: any, res: Response, _next: NextFunction): Promise<void> => {
     // get_products
-
     const resPerPage = 4;
-
     const productCount = await Product.countDocuments();
-
     const apiFeatures = new Api_Features(Product.find(), req.query)
       .search()
       .filter()
       .pagination(resPerPage);
-
     let products = await apiFeatures.query;
-
     const filteredProductsCount = products.length;
-
     products = await apiFeatures.query.clone();
-
     res.status(200).json({
       success: true,
       productCount,
@@ -47,11 +39,8 @@ export class Product_Controllers {
 
   public getSingleProduct = asyncError(async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     // get single product with id assigned by mongoose
-
     const product = await Product.findById(req.params.id);
-
     if (!product) return next(new Error_Handler('Product not found', 404));
-
     res.status(200).json({
       success: true,
       product,
@@ -60,14 +49,11 @@ export class Product_Controllers {
 
   public updateProduct = asyncError(async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     let product = await Product.findById(req.params.id);
-
     if (!product) return next(new Error_Handler('Product not found', 404));
-
     product = await Product.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
       runValidators: true,
     });
-
     res.status(200).json({
       success: true,
       product,
@@ -76,9 +62,7 @@ export class Product_Controllers {
 
   public deleteProduct = asyncError(async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     const product = await Product.findByIdAndRemove(req.params.id);
-
     if (!product) return next(new Error_Handler('Product not found', 404));
-
     res.status(200).json({
       success: true,
       message: 'Product deleted',
