+++ Backend/src/app/models/orders.ts
@@ -2,42 +2,8 @@ import mongoose, {
   Schema,
   Model,
   model,
-  Document,
 } from 'mongoose';
-import { IProducts } from './product';
-import { IUser } from './user';
-
-export interface IOrder extends Document {
-  shippingInfo: {
-    address: string,
-    city: string,
-    phoneNo: string,
-    postalCode: string,
-    country: string
-  },
-  user: IUser['_id'],
-  orderItems: orderItems[],
-  paymentInfo: {
-    id: string,
-    status: string,
-  }
-  paidAt: Date,
-  itemsPrice: number,
-  taxPrice: number,
-  shippingPrice: number,
-  totalPrice: number,
-  orderStatus: String,
-  deliveredAt: Date,
-  createdAt: Date
-}
-
-export interface orderItems extends IOrder {
-  name: string,
-  quantity: number,
-  image: string,
-  price: number,
-  product: IProducts['_id'],
-}
+import { IOrder } from '../interface/Iorder';
 
 const orderSchema: Schema = new Schema({
   shippingInfo: {
